// srcs/requirements/frontend/src/ui/avatarData/avatarData.types.ts
/**
 * @file avatarData.types.ts
 * @brief Defines the gypes used for avatar data in the application.
 */

/**
 * @interface avatar
 * @brief Represents an avatar with a name and an image path.
 *
 * @property {string} name - The name of the avatar.
 * @property {string} imagePath - The file path to the avatar's image.
 */

export interface avatar {
  name: string;
  imagePath: string;
}